changed datatypes from float to double

+++ processing/XOR_NeuralNetwork/Activation_Functions.pde
@@ -1,27 +1,27 @@
-private float dSigmoid(float x){
+private double dSigmoid(double x){
   return sigmoid(x) * (1 - sigmoid(x)); 
 }
 
-private float sigmoid(float x) {
-  return (float)(1/( 1 + Math.pow(Math.E,(-1*x))));
+private double sigmoid(double x) {
+  return (double)(1/( 1 + Math.pow(Math.E,(-1*x))));
 }
 
-private int bianary(float x){
-  float thesh = 0.5;
+private int bianary(double x){
+  double thesh = 0.5;
   if(x>=thesh)
     return 1;
   return 0;
 }
-private float linear(float x){
-  float slope =1;
-  float yint =0;
+private double linear(double x){
+  double slope =1;
+  double yint =0;
   
   return slope*x+yint;
 }
 
-private float relu(float x){
-  float thesh =0;
-  float lower =0;
+private double relu(double x){
+  double thesh =0;
+  double lower =0;
   if(x<thesh)
     return lower;
   return x;
